requirements.txt
requests
fastapi
python-dotenv
httpx
uvicorn
streamlit